# api/openapi-sistemas360.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Sistemas 360",
    "version": "1.0.0",
    "description": "API REST/JSON para facturación electrónica en Argentina. El entorno se determina por el token utilizado: los tokens de pruebas simulan la emisión y los tokens de producción autorizan comprobantes ante ARCA.\n\n**Facturación delegada — opción con activación previa**\n\nPodés emitir mediante nuestra API sin gestionar certificados ni claves privadas propios, autorizando a Sistemas 360 en ARCA. Para conocer las condiciones, solicitar la activación y recibir las instrucciones de integración, contactá a soporte@sistemas360.ar.",
    "termsOfService": "https://api.sistemas360.ar/terminos-y-condiciones",
    "contact": {
      "name": "Soporte API Sistemas 360",
      "email": "soporte@sistemas360.ar",
      "url": "https://api.sistemas360.ar/documentacion-api"
    }
  },
  "servers": [
    {
      "url": "https://api.sistemas360.ar",
      "description": "API Sistemas 360"
    }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/comprobantes/{comprobante}/reintentar": {
      "post": {
        "tags": ["Comprobantes"],
        "summary": "Reintentar comprobante",
        "description": "Sin body ni cabecera de escenario; requiere token del mismo entorno. Si ya está autorizado o completado en pruebas (simulado), devuelve 200 con el resultado existente, sin volver a emitir. En error reintenta la autorización conservando la fecha original, incluida la asignada al crear si se omitió fecha. Se revalidan las fechas: si ya no son válidas, devuelve 422 con fecha_reintento_invalida y mensaje_api, sin solicitar autorización ni modificar el registro. En producción ARCA valida el orden de fechas; pruebas lo controla antes de asignar número nuevo. Una preparación persistida en pendiente que no llegó a enviarse puede retomarse con los mismos datos y validación de fechas. En pendiente_confirmacion solo consulta la autorización original y verifica sus datos fiscales, aun si pasaron semanas, sin emitir otra vez; devuelve 409 si sigue sin confirmarse o recuperacion_inconsistente si no coincide. En pruebas usa el escenario guardado: pendiente_recuperable completa con estado simulado y pendiente_persistente mantiene 409; los errores técnicos recuperables completan si sus fechas siguen siendo válidas. Conservá el mismo ID y referencia mientras el resultado sea incierto. El integrador inicia los reintentos; no hay webhooks de estado. El resultado fiscal usa mensaje y error_type en el nivel principal, junto con data.estado; las validaciones locales pueden usar error.tipo y error.detalles.",
        "operationId": "retryReceipt",
        "parameters": [
          { "$ref": "#/components/parameters/ReceiptIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Comprobante autorizado, recuperado o ya completado anteriormente. No se duplica la emisión. Revisar data.estado y data.cae.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReceiptResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "Resultado pendiente de confirmación o datos incompatibles con el intento original (recuperacion_inconsistente). No se volvió a emitir; revisar mensaje_api.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReceiptResponse" } } }
          },
          "422": {
            "description": "Validación local, estado que no admite reintento o rechazo fiscal. fecha_reintento_invalida indica fechas originales no válidas para una nueva autorización; el registro permanece intacto. El rechazo fiscal usa mensaje y data.mensajes_arca.",
            "content": { "application/json": { "schema": { "anyOf": [
              { "$ref": "#/components/schemas/ErrorResponse" },
              { "$ref": "#/components/schemas/ReceiptResponse" }
            ] } } }
          },
          "503": {
            "description": "Servicio no disponible, respuesta incompleta u operación fiscal en curso. Conservar el comprobante y revisar data.estado y mensaje_api.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReceiptResponse" } } }
          },
          "504": {
            "description": "ARCA no respondio a tiempo. Revisar data.estado; puede seguir pendiente_confirmacion.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReceiptResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token API",
        "description": "Pegá únicamente el token, sin escribir Bearer. Podés generar un token de pruebas desde Panel → Primeros pasos. La credencial permanece solo en la memoria de esta pestaña."
      }
    },
    "parameters": {
...
      "ReceiptIdPath": {
        "name": "comprobante",
        "in": "path",
        "required": true,
        "description": "Identificador interno del comprobante.",
        "schema": { "type": "integer", "minimum": 1, "example": 59 }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Token ausente o inválido.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "ok": false,
              "error": { "codigo": 401, "mensaje": "Token inválido." }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Cuenta, token, empresa o entorno sin autorización.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "ok": false,
              "error": { "codigo": 403, "mensaje": "Token inactivo o suspendido." }
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso no encontrado para la cuenta.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "ok": false,
              "error": { "codigo": 404, "mensaje": "Recurso no encontrado." }
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiMessage": {
        "type": "string",
        "nullable": true,
        "description": "Explicación adicional de Sistemas 360 sobre el resultado y cómo continuar. Puede ser null. Texto para personas: automatizar usando HTTP, estado y códigos, no comparando frases. Los campos y mensajes originales se conservan.",
        "example": "El resultado está pendiente de confirmación. Esperá y usá el endpoint de reintento con el mismo ID."
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "mensaje_api": { "$ref": "#/components/schemas/ApiMessage" },
          "ok": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["codigo", "mensaje"],
            "properties": {
              "codigo": { "type": "integer", "example": 422 },
              "tipo": { "type": "string", "nullable": true, "example": "emisor_requerido" },
              "mensaje": { "type": "string", "example": "Datos inválidos." },
              "detalles": {
                "type": "object",
                "nullable": true,
                "additionalProperties": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          },
          "data": { "nullable": true }
        }
      },
...
      "ReceiptType": {
        "type": "string",
        "enum": [
          "factura_a", "factura_b", "factura_c",
          "nota_credito_a", "nota_credito_b", "nota_credito_c",
          "nota_debito_a", "nota_debito_b", "nota_debito_c"
        ]
      },
...
      "Tax": {
        "type": "object",
        "description": "Tributo adicional al IVA de los items, solo si corresponde segun el regimen fiscal, la operacion y la jurisdiccion. Confirmar los conceptos, bases y alicuotas con el responsable contable del emisor. El integrador envia los importes calculados; la API los incorpora al comprobante y valida el total.",
        "required": ["id", "base_imponible", "alicuota", "importe"],
        "properties": {
          "id": { "type": "integer", "minimum": 1, "description": "Codigo WSFE del tipo de tributo, no de la jurisdiccion. Habituales: 1 nacionales, 2 provinciales, 3 municipales, 4 internos, 6 percepcion IVA, 7 percepcion IIBB, 8 percepcion municipal, 99 otros tributos aceptados por ARCA. El 99 no admite cargos comerciales arbitrarios." },
          "descripcion": { "type": "string", "nullable": true, "maxLength": 255, "description": "Texto libre para identificar concepto y jurisdiccion. Para emitir ante ARCA, usar hasta 80 caracteres e informar siempre si id = 99. La validacion local admite hasta 255. Se informa a ARCA y aparece en respuesta y PDF." },
          "base_imponible": { "type": "number", "minimum": 0, "description": "Base imponible determinada por el integrador." },
          "alicuota": { "type": "number", "minimum": 0, "description": "Porcentaje: 2 significa 2 %. Enviar 0 si no corresponde una alicuota porcentual." },
          "importe": { "type": "number", "minimum": 0.01, "description": "Importe calculado por el integrador. Se suma al total; la API no lo calcula desde la base y alicuota." }
        },
        "example": { "id": 8, "descripcion": "Percepcion municipal - Municipio de ejemplo", "base_imponible": 10000, "alicuota": 2, "importe": 200 }
      },
...
      "ReceiptDetail": {
        "type": "object",
        "properties": {
          "codigo_tipo_comprobante": {"type":"integer","enum":[1,2,3,6,7,8,11,12,13,51,52,53],"description":"Código ARCA resuelto por la API. A común/CBU: 1,2,3; A retención: 51,52,53. No se envía al crear."},
          "modalidad_comprobantes_a": {"type":"string","nullable":true,"enum":["a_comun","a_cbu_informada","a_sujeta_retencion",null],"description":"Modalidad guardada al emitir. B/C y registros históricos sin modalidad identificable: null."},
          "id": { "type": "integer", "example": 59 },
          "empresa_id": { "type": "integer", "example": 9 },
          "empresa": { "type": "string", "nullable": true },
          "entorno": { "type": "string", "enum": ["pruebas", "produccion"] },
          "modo_simulacion": { "type": "boolean", "description": "true para todo comprobante de pruebas, también rechazados o pendientes. Es independiente del estado." },
          "estado": { "type": "string", "example": "autorizado" },
          "fecha_registro": { "type": "string", "nullable": true },
          "fecha_autorizacion": { "type": "string", "nullable": true },
          "tipo_comprobante": { "$ref": "#/components/schemas/ReceiptType" },
          "punto_venta": { "type": "integer" },
          "numero_comprobante": { "type": "integer" },
          "cae": { "type": "string", "nullable": true },
          "cae_vencimiento": { "type": "string", "format": "date", "nullable": true },
          "qr": {
            "type": "object",
            "properties": { "url": { "type": "string", "format": "uri", "nullable": true } }
          },
          "cliente": { "type": "object", "additionalProperties": true },
          "items": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "importes": {
            "type": "object",
            "properties": {
              "subtotal": { "type": "string" },
              "iva": { "type": "string" },
              "tributos": { "type": "string" },
              "total": { "type": "string" }
            }
          },
          "tributos": { "type": "array", "items": { "$ref": "#/components/schemas/Tax" } },
          "referencia_externa": { "type": "string", "nullable": true },
          "comprobante_asociado_id": { "type": "integer", "nullable": true },
          "comprobante_asociado": {"type":"object","nullable":true,"additionalProperties":true,"properties":{"codigo_tipo_comprobante":{"type":"integer","enum":[1,2,3,6,7,8,11,12,13,51,52,53],"description":"Código ARCA resuelto por la API. A común/CBU: 1,2,3; A retención: 51,52,53. No se envía al crear."},"modalidad_comprobantes_a":{"type":"string","nullable":true,"enum":["a_comun","a_cbu_informada","a_sujeta_retencion",null],"description":"Modalidad guardada al emitir. B/C y registros históricos sin modalidad identificable: null."}}},
          "mensajes_arca": { "type": "array", "description": "Mensajes originales recibidos de ARCA. En escenarios de pruebas permanece vacío; consultar mensaje_api.", "items": { "type": "object", "additionalProperties": true } },
          "imprimir_a4_url": { "type": "string", "format": "uri" },
          "imprimir_ticket_url": { "type": "string", "format": "uri" }
        }
      },
      "ReceiptResponse": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "mensaje_api": { "$ref": "#/components/schemas/ApiMessage" },
          "ok": { "type": "boolean", "example": true },
          "mensaje": { "type": "string", "nullable": true },
          "error_type": { "type": "string", "nullable": true, "description": "Tipo de error devuelto por el endpoint de reintento, por ejemplo pendiente_confirmacion o arca_timeout." },
          "data": { "$ref": "#/components/schemas/ReceiptDetail" }
        }
      },
...
    }
  }
}